refactor: organizar base em domínios e consolidar utilitários #752
