+++ apps/apae/src/app/person/register/responsible/page.tsx
@@ -19,7 +19,7 @@ import { EditGuardian } from "@/schemas/edit-member-schemas";
 import { zodResolver } from "@hookform/resolvers/zod";
 import React, { useState, useEffect } from "react"; 
 import { useForm } from "react-hook-form";
-import { handleBackendValidationErrors } from "@/utils/form-errors";
+import { handleBackendValidationErrors } from "@/lib/utils/form-errors";
 import { usePathname } from "next/navigation"; 
 import { formatPhone } from "@/lib/formats";
 import { Checkbox } from "@/components/ui/checkbox";
