+++ apps/apae/src/lib/utils/form-errors.ts
@@ -1,8 +1,5 @@
-import { UseFormSetError, FieldValues, Path } from "react-hook-form";
+import { FieldValues, Path, UseFormSetError } from "react-hook-form";
 
-/**
- * Interface que espelha o ValidationErrorResponse do Java.
- */
 interface ApiFieldError {
   field: string;
   message: string;
@@ -13,12 +10,6 @@ interface ApiErrorResponse {
   message?: string;
 }
 
-/**
- * Mapeia erros de validação do backend para os campos do formulário no frontend.
- * Resolve nomes de campos aninhados e mapeia campos específicos de API para a UI.
- * * @param errorData Objeto de erro retornado pela API contendo a lista de campos.
- * @param setError Função setError proveniente do useForm do React Hook Form.
- */
 export function handleBackendValidationErrors<T extends FieldValues>(
   errorData: ApiErrorResponse,
   setError: UseFormSetError<T>,
